--- components/styles.js
@@ -1,22 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-// Layaout
-
-const styles = StyleSheet.create({
-  screen: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-});
-
-// WelcomeScreen
-// LoginScreen
-// HomeScreen
-
-const saldoText = StyleSheet.create({
-  fontSize: 30,
-  padding: 20,
-});
-
-export { styles, saldoText };
