+++ screens/AuthScreen/styles.js
@@ -1,0 +1,4 @@
+import crossStyles from '../../styles/CrossStyles';
+
+const styles = crossStyles;
+export default styles;
