+++ screens/HomeScreen/HomeScreen.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { View, Text } from 'react-native';
 import { useSelector } from 'react-redux';
 import { Avatar } from 'react-native-elements';
-import { styles, saldoText } from '../../components/styles';
+import styles from './styles';
 
 function HomeScreen() {
   const email = useSelector(state => state.auth.user.email);
@@ -20,7 +20,7 @@ function HomeScreen() {
       />
       <Text>{`Hola ${email}!`}</Text>
 
-      <Text style={saldoText}>{`Saldo ${saldo} `}</Text>
+      <Text style={styles.saldoText}>{`Saldo ${saldo} `}</Text>
 
     </View>
