+++ screens/WelcomeScreen/styles.js
@@ -1,0 +1,11 @@
+import { StyleSheet } from 'react-native';
+import crossStyles from '../../styles/CrossStyles';
+
+const styles = StyleSheet.create({
+  saldoText: {
+    fontSize: 30,
+    padding: 20,
+  },
+});
+
+export default { ...crossStyles, ...styles };
