+++ styles/CrossStyles.js
@@ -1,0 +1,11 @@
+import { StyleSheet } from 'react-native';
+
+const crossStyles = StyleSheet.create({
+  screen: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+});
+
+export default crossStyles;
